+++ src/server/routes/general/project-information.ts
@@ -1,3 +1,4 @@
+import { getVersion } from '../../../visreg';
 import { getSuites } from '../../../utils';
 const express = require('express');
 const router = express.Router();
@@ -5,8 +6,11 @@ const router = express.Router();
 const suites = getSuites();
 
 router.get('/project-information', (req: any, res: any) => {
+    const version = getVersion();
+    
     res.send({
-        suites
+        suites,
+        version,
     });
 });