+++ web/src/pages/assessment-page/assessment-page.tsx
@@ -3,22 +3,70 @@ import { AppContext } from '../../contexts/app-context';
 import Controls from './controls';
 import Progress from './progress';
 import { useLoaderData, useNavigate } from 'react-router-dom';
-import { AssessmentData } from './types';
+import { AssessmentData, DiffObject } from './types';
 import { getImageDetails } from '../../loaders-and-fetchers';
 import PreviewComponent from '../../components/image-viewer/preview-component';
 import { Image } from '../../types';
+import { PrevNextButton } from '../../components/ui/prev-next-button';
+import x from '@stylexjs/stylex';
+import { useTheme } from '@mui/material';
+
+type TemporaryAssessmentResultsLookupTable = {
+	[index: number]: DiffObject;
+};
+
+const s = x.create({
+	approvedLightmode: {
+		backgroundColor: 'rgb(243, 247, 243)',
+		border: '1px solid rgba(143, 175, 144, 0.8)',
+		borderRadius: '0.75rem',
+	},
+	approvedDarkmode: {
+		backgroundColor: 'rgba(243, 255, 243, 0.08)',
+		border: '1px solid rgba(143, 175, 144, 0.6)',
+		borderRadius: '0.75rem',
+	},
+	rejectedLightmode: {
+		backgroundColor: 'rgba(250, 236, 237, 1)',
+		border: '1px solid rgba(211, 74, 80, 0.8)',
+		borderRadius: '0.5rem',
+	},
+	rejectedDarkmode: {
+		backgroundColor: 'rgba(255, 94, 109, 0.08)',
+		border: '1px solid rgba(211, 74, 80, 0.6)',
+		borderRadius: '0.5rem',
+	},
+	assessmentPage: {
+
+	},
+	controlsContainer: {
+	},
+	prevNextContainer: {
+        width: 'auto',
+        display: 'flex',
+        textOverflow: 'ellipsis',
+        justifyContent: 'space-between',
+        gap: '1rem',
+        maxWidth: '650px',
+        margin: '0 auto',
+        marginBlock: '1rem',
+    },
+});
 
 
 const AssessmentPage = () => {
 	const [ fetchedImagesDetails, setFetchedImagesDetails ] = useState<Image[]>([]);
+	const [ temporaryAssessmentData, setTemporaryAssessmentData ] = useState<TemporaryAssessmentResultsLookupTable>([]);
 	const { api, setSuiteName, setCurrentDiffIndex, currentDiffIndex } = useContext(AppContext);
 	const { assessmentData } = useLoaderData() as { assessmentData: AssessmentData; };
 	const navigate = useNavigate();
+	const theme = useTheme();
 
 	useEffect(() => {
 		if (!assessmentData) return;
 
 		const diffFiles = assessmentData?.diffFiles;
+		setTemporaryAssessmentData(assessmentData.temporaryAssessmentResults);
 
 		if (!diffFiles) {
 			setCurrentDiffIndex(null);
@@ -63,11 +111,14 @@ const AssessmentPage = () => {
 	const doAssessAction = async (action: string) => {
 		if (!assessmentData) return;
 
-		await fetch(api + '/assessment/' + action, {
+		const res = await fetch(api + '/assessment/' + action, {
 			method: 'POST',
 			headers: { 'Content-Type': 'application/json' },
 			body: JSON.stringify({ diffImage: assessmentData.diffFiles[currentDiffIndex!] }),
 		});
+
+		const tempAssessmentResults: TemporaryAssessmentResultsLookupTable = await res.json();
+		setTemporaryAssessmentData(tempAssessmentResults);
 
 		if (currentDiffIndex === assessmentData.diffFiles.length - 1) {
 			navigate('/summary');
@@ -77,15 +128,58 @@ const AssessmentPage = () => {
 		if (currentDiffIndex !== null) {
 			setCurrentDiffIndex(currentDiffIndex + 1);
 		}
-	};		
+	};
+
+	const prevNextClickHandler = (direction: string) => {
+		if (currentDiffIndex === null) return;
+		const newIndex = direction === 'next' ? currentDiffIndex + 1 : currentDiffIndex - 1;
+		setCurrentDiffIndex(newIndex)
+	};
 
 	return (
 		<>
-			{currentDiffIndex !== null && fetchedImagesDetails[currentDiffIndex] && (
-				<PreviewComponent image={fetchedImagesDetails[currentDiffIndex]}>
-					<Controls doAssessAction={(action: string) => doAssessAction(action)} />
-					<Progress currentDiffIndex={currentDiffIndex} diffFiles={assessmentData.diffFiles} />
-				</PreviewComponent>
+			{currentDiffIndex !== null && (
+				<div {...x.props(
+					s.assessmentPage,
+					theme.palette.mode === 'dark'
+						? temporaryAssessmentData[ currentDiffIndex ]?.assessedAs === 'rejected' && s.rejectedDarkmode
+						: temporaryAssessmentData[ currentDiffIndex ]?.assessedAs === 'rejected' && s.rejectedLightmode,
+					theme.palette.mode === 'dark'
+						? temporaryAssessmentData[ currentDiffIndex ]?.assessedAs === 'approved' && s.approvedDarkmode
+						: temporaryAssessmentData[ currentDiffIndex ]?.assessedAs === 'approved' && s.approvedLightmode,
+				)}>
+			
+					{fetchedImagesDetails[ currentDiffIndex ] && (
+						<PreviewComponent image={fetchedImagesDetails[ currentDiffIndex ]}>
+							<div {...x.props(s.controlsContainer)}>
+
+								<div {...x.props(s.prevNextContainer)}>
+									{currentDiffIndex - 1 >= 0 ? (
+										<PrevNextButton
+											title={currentDiffIndex ? assessmentData.diffFiles[ currentDiffIndex - 1 ].imageName : ''}
+											direction='prev'
+											clickHandler={() => prevNextClickHandler('prev')}
+										/>
+									) : <div />}
+
+									{(currentDiffIndex + 1 < assessmentData.diffFiles.length) &&
+										temporaryAssessmentData[ currentDiffIndex]?.assessedAs ? (
+										<PrevNextButton
+											title={assessmentData.diffFiles[ currentDiffIndex + 1 ].imageName}
+											direction='next'
+											clickHandler={() => prevNextClickHandler('next')}
+										/>
+									) : <div />}
+								</div>
+
+								<Controls doAssessAction={(action: string) => doAssessAction(action)} />
+
+							</div>
+							<Progress currentDiffIndex={currentDiffIndex} diffFiles={assessmentData.diffFiles} />
+						</PreviewComponent>
+					)}
+				</div>
+
 			)}
 		</>
 	);