+++ web/src/pages/suite-page/suite-home/suite-home.tsx
@@ -72,7 +72,7 @@ const SuiteHome = () => {
                 </Link>
             </Typography>
 
-            <Actions suiteSlug={suiteSlug} />
+            <Actions />
 
             <div {...x.props(s.infoArea)}>
                 <SuiteDetailsSidebar />